# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Memberslist API",
    "version": "1.0.0",
    "description": "구글 스프레드시트를 기반으로 한 회원/주문/메모/후원수당 관리 API"
  },
  "servers": [
    { "url": "https://memberslist.onrender.com" }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "홈 (Health Check)",
        "description": "서버 정상 실행 여부 확인",
        "operationId": "getHealthCheck",
        "responses": {
          "200": { "description": "Flask 서버 실행 메시지" }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "OpenAPI 스펙",
        "description": "현재 API의 OpenAPI JSON 스펙 반환",
        "operationId": "getOpenapiSpec",
        "responses": {
          "200": { "description": "OpenAPI JSON 스펙 반환" }
        }
      }
    },
    "/debug_sheets": {
      "get": {
        "summary": "구글 시트 디버깅",
        "description": "현재 연결된 Google Sheets 목록과 특정 시트의 헤더 확인",
        "operationId": "getDebugSheets",
        "parameters": [
          {
            "name": "sheet",
            "in": "query",
            "description": "조회할 시트명 (예: DB, 제품주문, 상담일지)",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "시트 목록 또는 헤더 반환" },
          "500": { "description": "시트 조회 오류" }
        }
      }
    },
    "/postIntent": {
      "post": {
        "summary": "자연어 intent 처리",
        "description": "자연어 입력을 intent와 query dict로 변환하여 실행",
        "operationId": "postIntent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NLURequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "intent 처리 성공" },
          "400": { "description": "요청 오류" }
        }
      }
    },
    "/guess_intent": {
      "post": {
        "summary": "intent 실행 게이트웨이",
        "description": "nlu_to_pc_input에서 intent를 추출하고 INTENT_MAP을 통해 실행",
        "operationId": "guessIntentEntry",
        "responses": {
          "200": { "description": "intent 실행 성공" },
          "400": { "description": "지원하지 않는 intent" }
        }
      }
    },
    "/member": {
      "post": {
        "summary": "회원 관련 API",
        "description": "회원 intent 기반 단일 라우트 (조회/등록/수정/삭제)",
        "operationId": "postMember",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MemberRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "회원 관련 처리 성공" },
          "400": { "description": "요청 오류" }
        }
      }
    },
    "/order": {
      "post": {
        "summary": "주문 관련 API",
        "description": "제품 주문 intent 기반 단일 라우트",
        "operationId": "postOrder",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderRequest" }
            },
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/OrderFileUpload" }
            }
          }
        },
        "responses": {
          "200": { "description": "주문 처리 성공" },
          "400": { "description": "요청 오류" }
        }
      }
    },
    "/memo": {
      "post": {
        "summary": "메모 관련 API",
        "description": "상담일지/개인메모 intent 기반 단일 라우트",
        "operationId": "postMemo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MemoRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "메모 처리 성공" },
          "400": { "description": "요청 오류" }
        }
      }
    },
    "/commission": {
      "post": {
        "summary": "후원수당 관련 API",
        "description": "후원수당 intent 기반 단일 라우트",
        "operationId": "postCommission",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CommissionRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "후원수당 처리 성공" },
          "400": { "description": "요청 오류" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "NLURequest": {
        "type": "object",
        "properties": {
          "text": { "type": "string" },
          "query": { "type": "string" }
        },
        "minProperties": 1
      },
      "MemberRequest": {
        "type": "object",
        "oneOf": [
          {
            "properties": { "text": { "type": "string" } },
            "required": ["text"]
          },
          {
            "properties": {
              "회원명": { "type": "string" },
              "회원번호": { "type": "string" },
              "휴대폰번호": { "type": "string" }
            }
          }
        ]
      },
      "OrderRequest": {
        "type": "object",
        "oneOf": [
          {
            "properties": { "text": { "type": "string" } },
            "required": ["text"]
          },
          {
            "properties": {
              "주문회원": { "type": "string" },
              "상품": { "type": "string" },
              "수량": { "type": "integer" },
              "결제": { "type": "string" },
              "date": { "type": "string", "format": "date" }
            }
          }
        ]
      },
      "OrderFileUpload": {
        "type": "object",
        "properties": {
          "file": { "type": "string", "format": "binary" }
        }
      },
      "MemoRequest": {
        "type": "object",
        "oneOf": [
          {
            "properties": { "text": { "type": "string" } },
            "required": ["text"]
          },
          {
            "properties": {
              "회원명": { "type": "string" },
              "내용": { "type": "string" },
              "일지종류": { "type": "string" }
            },
            "required": ["회원명", "내용", "일지종류"]
          }
        ]
      },
      "CommissionRequest": {
        "type": "object",
        "oneOf": [
          {
            "properties": { "text": { "type": "string" } },
            "required": ["text"]
          },
          {
            "properties": {
              "회원명": { "type": "string" },
              "기간": { "type": "string" },
              "월": { "type": "string" }
            }
          }
        ]
      }
    }
  }
}